ProductWithAmount abstraction for recipes and inventories
